+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   - Add support for feedback survey responses
   - Change all instances of ADD to COPY in Dockerfile
   - Remove use of SDX_HOME variable in makefile
+  - Fix Postgres DB tests
 
 ### 2.0.0 2017-07-11
   - Version bump